+++ components/shared/close-dialog.tsx
@@ -1,19 +1,15 @@
 import { IoMdClose } from "react-icons/io";
 import store from "@/store";
+import { cn } from "@/lib/utils";
 
-export default function CloseModal() {
-  const { closeDialog, closeDrawer } = store();
-
-  const handleClose = () => {
-    closeDialog();
-    closeDrawer();
-  };
+export default function CloseDialog({ className }: { className?: string }) {
+  const { closeDialog } = store();
 
   return (
     <button
       type="button"
-      onClick={handleClose}
-      className="absolute right-4 top-4"
+      onClick={closeDialog}
+      className={cn("absolute right-4 top-4", className)}
     >
       <IoMdClose size={20} />
     </button>